backend/requirements.txt
asgiref==3.11.0
async-timeout==5.0.1
blinker==1.9.0
click==8.3.1
Django==5.2.9
djangorestframework==3.16.1
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
peewee==3.18.3
Pygments==2.19.2
redis==5.0.1
sqlite-web==0.6.6
sqlparse==0.5.5
typing_extensions==4.15.0
Werkzeug==3.1.4
